+++ src/flexpath.h
@@ -82,11 +82,12 @@ struct FlexPath {
     // Otherwise, num_elements and elements are expected to be already
     // allocated and set.  Arguments width, offset and tag can be single values
     // (which are applied to all elements) or arrays with count num_elements,
-    // one value for each path element.
+    // one value for each path element.  Argument separation is the desired
+    // distance between adjacent elements.
     void init(const Vec2 initial_position, double width, double offset, double tolerance, Tag tag);
     void init(const Vec2 initial_position, const double* width, const double* offset,
               double tolerance, const Tag* tag);
-    void init(const Vec2 initial_position, uint64_t num_elements_, double width, double offset,
+    void init(const Vec2 initial_position, uint64_t num_elements_, double width, double separation,
               double tolerance, Tag tag);
     void init(const Vec2 initial_position, uint64_t num_elements_, const double* width,
               const double* offset, double tolerance, const Tag* tag);